// frontend/spot-fix/src/app/shared/api/models/LocationDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spot Fix API
 * Spot Fix API documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WardDTO } from './WardDTO';
import {
    WardDTOFromJSON,
    WardDTOFromJSONTyped,
    WardDTOToJSON,
    WardDTOToJSONTyped,
} from './WardDTO';

/**
 * 
 * @export
 * @interface LocationDTO
 */
export interface LocationDTO {
    /**
     * 
     * @type {number}
     * @memberof LocationDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof LocationDTO
     */
    addressText: string;
    /**
     * 
     * @type {string}
     * @memberof LocationDTO
     */
    landmark?: string;
    /**
     * 
     * @type {number}
     * @memberof LocationDTO
     */
    latitude?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationDTO
     */
    longitude?: number;
    /**
     * 
     * @type {WardDTO}
     * @memberof LocationDTO
     */
    ward?: WardDTO;
}

/**
 * Check if a given object implements the LocationDTO interface.
 */
export function instanceOfLocationDTO(value: object): value is LocationDTO {
    if (!('addressText' in value) || value['addressText'] === undefined) return false;
    return true;
}

export function LocationDTOFromJSON(json: any): LocationDTO {
    return LocationDTOFromJSONTyped(json, false);
}

export function LocationDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): LocationDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'addressText': json['addressText'],
        'landmark': json['landmark'] == null ? undefined : json['landmark'],
        'latitude': json['latitude'] == null ? undefined : json['latitude'],
        'longitude': json['longitude'] == null ? undefined : json['longitude'],
        'ward': json['ward'] == null ? undefined : WardDTOFromJSON(json['ward']),
    };
}

export function LocationDTOToJSON(json: any): LocationDTO {
    return LocationDTOToJSONTyped(json, false);
}

export function LocationDTOToJSONTyped(value?: LocationDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'addressText': value['addressText'],
        'landmark': value['landmark'],
        'latitude': value['latitude'],
        'longitude': value['longitude'],
        'ward': WardDTOToJSON(value['ward']),
    };
}